// impartner-prm-services/client/packages/web-widgets/src/app/core/widget-tag.ts
export enum WidgetTag {
  HelloWorld = 'hello-world',
  HelloCustomer = 'hello-customer',
  RichTextEditorEdit = 'rich-text-editor-edit',
  RichTextEditorView = 'rich-text-editor-view',
  HtmlEditorEdit = 'html-editor-edit',
  HtmlEditorView = 'html-editor-view',
  Grid = 'grid',
  PrmObjectGridEdit = 'prm-object-grid-edit',
  PrmObjectGridView = 'prm-object-grid-view',
  PrmObjectGridSettings = 'prm-object-grid-settings',
  SpacerEdit = 'spacer-edit',
  SpacerView = 'spacer-view',
  SpacerSettings = 'spacer-settings-popover',
  GridFiltersEdit = 'grid-filters-view',
  GridFiltersView = 'grid-filters-edit',
  GridFiltersSettings = 'grid-filters-settings'
}